Merge index.html changes

+++ pyth.py
@@ -13,22 +13,23 @@ import random
 import os
 
 project_dir = os.path.dirname(os.path.abspath(__file__))
-database_file = "sqlite:///{}".format("links.db")
 
-app = Flask(__name__)
-app.config["SQLALCHEMY_DATABASE_URI"] = database_file
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+application = Flask(__name__)
+application.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get('DATABASE_URL') or 'sqlite:///' + os.path.join(project_dir, 'links.db')
+application.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
 
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
+db = SQLAlchemy(application)
+migrate = Migrate(application, db)
 
 class Link(db.Model):
-    link = db.Column(db.String(80), unique=True, nullable=False, primary_key=True)
-    target = db.Column(db.String(80), nullable=False)
-    password = db.Column(db.String(80))
+    link = db.Column(db.Text(), unique=True, nullable=False, primary_key=True)
+    target = db.Column(db.Text(), nullable=False)
+    password = db.Column(db.Text(), nullable=False)
 
     def __repr__(self):
         return "<Link: {}>".format(self.link)
+
+db.create_all()
 
 illegal_links = ["__repl", "peek", "decoder", "toggle-peek", "make", "delete", "change-link", "change-target", "decode"]
 
@@ -44,7 +45,7 @@ def make_url():
             link = ''.join(random.choices(chars, k=length))
             chosen = link not in tried
         
-        tried.append(link)
+        tried.applicationend(link)
         if Link.query.filter(Link.link == tried[-1]).first():
             length += 1
             tried = []
@@ -53,11 +54,11 @@ def make_url():
     
     return tried[-1]
 
-@app.shell_context_processor
+@application.shell_context_processor
 def make_shell_context():
     return {'db': db, 'Link': Link}
 
-@app.route('/<page>')
+@application.route('/<page>')
 def go_to_page(page):
     if request.cookies.get("peek") == "1":
         return redirect("peek/" + page)
@@ -68,7 +69,7 @@ def go_to_page(page):
         except:
             return render_template("error.html") 
 
-@app.route('/<page>/<extra>')
+@application.route('/<page>/<extra>')
 def go_to_page_extra(page, extra):
     if request.cookies.get("peek") == "1":
         return redirect("/peek/" + page + "/" + extra)
@@ -83,7 +84,7 @@ def go_to_page_extra(page, extra):
         except:
             return render_template("error.html") 
 
-@app.route('/peek/<page>')
+@application.route('/peek/<page>')
 def peek_page(page):
     try:
         target = Link.query.get(page).target
@@ -91,7 +92,7 @@ def peek_page(page):
     except:
         return render_template("error.html")
 
-@app.route('/peek/<page>/<extra>')
+@application.route('/peek/<page>/<extra>')
 def peek_page_extra(page, extra):
     try:
         target = Link.query.get(page).target
@@ -103,15 +104,15 @@ def peek_page_extra(page, extra):
     except:
         return render_template("error.html") 
 
-@app.route('/')
+@application.route('/')
 def index():
 	return render_template('index.html', peek="On" if request.cookies.get("peek") == "1" else "Off")
 
-@app.route('/decoder')
+@application.route('/decoder')
 def decode():
     return render_template('decode.html')
 
-@app.route('/toggle-peek')
+@application.route('/toggle-peek')
 def toggle_peek():
     resp = make_response(redirect("https://www.pyth.link"))
     if request.cookies.get("peek") in (None, "0"):
@@ -120,7 +121,7 @@ def toggle_peek():
         resp.set_cookie("peek", "0")
     return resp
 
-@app.route('/make', defaults={'target': None, 'link': None, 'password': None})
+@application.route('/make', defaults={'target': None, 'link': None, 'password': None})
 def make_link(target, link, password):  
     if link == None:
         link = request.args.get('link') 
@@ -132,7 +133,7 @@ def make_link(target, link, password):
         target = "http://" + target
     if not ((link.isalnum() or not link) and password and checkers.is_url(target)):
         return render_template("input_error.html")
-    if Link.query.filter(Link.link == link).first() or link in illegal_links:
+    if Link.query.get(link) or link in illegal_links:
         return render_template("taken.html")
     else:
         if not link:
@@ -142,7 +143,7 @@ def make_link(target, link, password):
         db.session.commit()
         return render_template("created.html", link=link, password=password)
 
-@app.route('/delete', defaults={'link': None, 'password': None})
+@application.route('/delete', defaults={'link': None, 'password': None})
 def delete_link(link, password):   
     if link == None:
         link = request.args.get('link')
@@ -150,18 +151,19 @@ def delete_link(link, password):
         password = request.args.get('password')
     if not (link and password):
         return render_template("input_error.html")
-    try:
-        password_real = Link.query.get(link).password
+    deleted_link = Link.query.get(link)
+    if deleted_link:
+        password_real = deleted_link.password
         if password == str(password_real):
             Link.query.filter_by(link=link).delete()
             db.session.commit()
             return render_template("removed.html")
         else:
             return render_template("wrong.html")
-    except:
+    else:
         return render_template("error.html")
 
-@app.route('/change-link', defaults={'link': None, 'password': None, 'new_link': None})
+@application.route('/change-link', defaults={'link': None, 'password': None, 'new_link': None})
 def change_link(link, password, new_link):   
     if link == None:
         link = request.args.get('link')
@@ -180,12 +182,12 @@ def change_link(link, password, new_link):
             return render_template("changed.html")
         elif password != str(password_real):
             return render_template("wrong.html")
-        elif Link.query.filter(Link.link == new_link).first() or link in illegal_links:
+        elif Link.query.get(link) or link in illegal_links:
             return render_template("taken.html")
     except:
         return render_template("error.html")
 
-@app.route('/change-target', defaults={'link': None, 'password': None, 'new_target': None})
+@application.route('/change-target', defaults={'link': None, 'password': None, 'new_target': None})
 def change_target(link, password, new_target):   
     if link == None:
         link = request.args.get('link')
@@ -207,7 +209,7 @@ def change_target(link, password, new_target):
     except:
         return render_template("error.html")
 
-@app.route('/decode', defaults={'link': None})
+@application.route('/decode', defaults={'link': None})
 def decode_link(link):
     try:
         if link == None:
@@ -220,4 +222,4 @@ def decode_link(link):
         return render_template("error.html")
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=8080)
+    application.run(host='0.0.0.0', port=8080)
